Move libraryEditor to the same level with libSandbox.
