+++ src/app/libraryEditor/components/Backdrop.js
