+++ src/app/libraryEditor/components/DataElementEdit.js
@@ -31,7 +31,7 @@ import {
 
 import ComponentIframe from './ComponentIframe';
 import Backdrop from './Backdrop';
-import NAMED_ROUTES from '../../../constants';
+import NAMED_ROUTES from '../../constants';
 
 const isNewComponent = ({
   match: {
