+++ src/app/libraryEditor/components/ExtensionConfigurationEdit.js
@@ -17,7 +17,7 @@ import { Map, List } from 'immutable';
 import { Flex, View, Heading, Picker, Item, ButtonGroup, Button } from '@adobe/react-spectrum';
 import ComponentIframe from './ComponentIframe';
 import Backdrop from './Backdrop';
-import NAMED_ROUTES from '../../../constants';
+import NAMED_ROUTES from '../../constants';
 
 const isNewExtensionConfiguration = ({
   extensionConfigurations,
