+++ src/app/libraryEditor/components/ExtensionConfigurationsList.js
