+++ src/app/libraryEditor/components/Settings.css
