+++ src/app/libraryEditor/components/Settings.js
