+++ src/app/libraryEditor/models/company.js
