+++ src/app/libraryEditor/models/currentRule.js
