+++ src/app/libraryEditor/models/extensions.js
