+++ src/app/libraryEditor/models/otherSettings.js
