+++ src/app/libraryEditor/models/property.js
