+++ src/app/libraryEditor/models/registry.js
